Реализована загрузка и сохранение данных в БД

+++ EquipmentCalculating/EquipmentCalculating/Models/Entities/Category.cs
@@ -2,7 +2,7 @@
 {
     public class Category
     {
-        internal int Id { get; set; }
+        internal long Id { get; set; }
 
         public string Name { get; set; }
